requirements.txt
sounddevice
numpy
openai-whisper
edge-tts
pygame
PyQt5
opencv-python
pyautogui
google-genai
